+++ src/explorer/models/kafka.ts
@@ -1,7 +1,7 @@
 import { Disposable, TreeItemCollapsibleState } from "vscode";
 import { ClientAccessor } from "../../client";
 import { ClusterSettings } from "../../settings";
-import { ClusterItem, NoClusterItem } from "./cluster";
+import { ClusterItem} from "./cluster";
 import { NodeBase } from "./nodeBase";
 
 export class KafkaModel extends NodeBase implements Disposable {
@@ -17,9 +17,6 @@ export class KafkaModel extends NodeBase implements Disposable {
 
     public async computeChildren(): Promise<NodeBase[]> {
         const clusters = this.clusterSettings.getAll();
-        if (clusters.length === 0) {
-            return [new NoClusterItem(this)];
-        }
         return clusters.map((c) => {
             return new ClusterItem(this.clientAccessor.get(c.id), c, this);
         });